// apps/api/src/modules/buyback/entities/buyback-lead.entity.d.ts
import { BuybackPhoto } from './buyback-photo.entity';
export declare class BuybackLead {
    photos: BuybackPhoto[];
    id: string;
    brand: string;
    modelName: string;
    phone: string;
    deviceType: string;
    status: string;
    screenCondition: string | null;
    bodyCondition: string | null;
    batteryHealth: string | null;
    functionalIssues: string | null;
    notes: string | null;
    createdAt: Date;
}
//# sourceMappingURL=buyback-lead.entity.d.ts.map